Move `r_new_list()` out of the C library
